Add Checkboxes markup
This commit adds the controlling checkboxes to the markup in an
unordered list and nests the related checkboxes to indicate which
checkboxes are linked. As the framework will depend on the `kjs-id` and
`kjs-checkbox-id` attributes, the checkboxes should function regardless
of their position on the page.

I'm conscious of the fact that the attributes used are quite repetitive
for "IDs". This was thought to be in keeping with the earlier modules
defined in the framework which do allow for duplicate values in the
markup. Hopefully this choice won't be too controversial.